Fixed machine-gui desync after auto reset
From the uC source, the only change was removing the debug
condition from the ON_RESET report, now the uC sends it to the
serial port, always. A "!" was added to the end of that print to
adhere to the comm standard.
Finally, a .gitignore file was added to facilitate the following
commits.